Merge component, Git support for three-way merge of block markup documents (#3)
## Description

Enables the `static-files-editor` plugin to gracefully merge files
changed both by the local app and a remote app.

This PR came out much larger than I expected. I'll still try to get away
with a rather short description as I consider this repository to be
still in flux before what we'll call the first consumable version. Think
of this PR as a of a part of the initial commit that still provides a
little bit of context completely for free :-)

The approach is as follows:

1. Edit a local file in the block editor and send the update to the
server
2. If the local file is different from the `post_content` database
field, run a three way merge where `post_content` is the parent, local
file is branch A, and the changes from the editor are branch B.
3. Send the updated markup to the browser
4. If the server responds with a different markup than it was given,
JavaScript runs a three-way merge where the submitted markup is the
parent, server markup is branch A, and the latest editor markup is
branch B.

The three-way merge works as follows:

1. Compute a diff from (P)arent to A
2. Compute a diff from (P)arent to B
3. Convert diffs PA and PB to chunks like `$base_text, $was_deleted,
$insertion`
4. Align PA and PB chunks to have pairs of chunks targeting the same
$base_text
5. Run through a set of heuristics to produce a merged version, e.g.
"both deleted" = delete, "both inserted" = conflict
6. On clean merge, move forward with the merged document
7. On conflict, store version A in history and use version B as the new
document

## Other important changes

This PR also furthers the Git library, here are the most important
changes:

### Merge – new module

Support for three-way merges with different diffing, merging, and
validation strategies:

```php
$strategy = new MergeStrategy(
	new MyersDiffer(),
	new ChunkMerger(),
	new BlockMarkupMergeValidator()
);
$merge_result = $strategy->merge( $parent, $changeA, $changeB )
```

### Git

New methods:

* `$git_remote->pull( $full_branch_name, $options )`
* `$git_repository->merge( $branch_name, [ 'merge_function' => function(
$parent, $a, $b) {} ] )`